:recycle: api: bye flash-certification context
