+++ api/server.js
@@ -18,7 +18,6 @@ import {
 } from './src/certification/configuration/routes.js';
 import { certificationEnrolmentRoutes } from './src/certification/enrolment/routes.js';
 import { certificationEvaluationRoutes } from './src/certification/evaluation/routes.js';
-import { flashCertificationRoutes } from './src/certification/flash-certification/routes.js';
 import { certificationResultRoutes } from './src/certification/results/routes.js';
 import { scoringRoutes } from './src/certification/scoring/routes.js';
 import { certificationSessionRoutes } from './src/certification/session-management/routes.js';
@@ -235,7 +234,6 @@ const setupRoutesAndPlugins = async function (server) {
     profileRoutes,
     questRoutes,
     evaluationRoutes,
-    flashCertificationRoutes,
     devcompRoutes,
     schoolRoutes,
     teamRoutes,
