+++ api/src/certification/evaluation/application/scenario-simulator-route.js
