+++ api/src/certification/evaluation/domain/models/AssessmentSimulatorSingleMeasureStrategy.js
