+++ api/src/certification/evaluation/domain/models/FlashAssessmentAlgorithmChallengesBetweenCompetencesRule.js
