+++ api/src/certification/evaluation/domain/services/algorithm-methods/flash.js
