+++ api/src/certification/evaluation/infrastructure/serializers/scenario-simulator-batch-serializer.js
