+++ api/src/certification/evaluation/routes.js
@@ -1,13 +1,15 @@
-import * as companionAlert from '../evaluation/application/companion-alert-route.js';
+import * as scenarioSimulatorRoutes from '../evaluation/application/scenario-simulator-route.js';
 import * as certificationAdminRoutes from './application/certification-admin-route.js';
 import * as certificationCourseRoutes from './application/certification-course-route.js';
 import * as certificationRescoringRoutes from './application/certification-rescoring-route.js';
+import * as companionAlert from './application/companion-alert-route.js';
 
 const certificationEvaluationRoutes = [
   companionAlert,
   certificationCourseRoutes,
   certificationAdminRoutes,
   certificationRescoringRoutes,
+  scenarioSimulatorRoutes,
 ];
 
 export { certificationEvaluationRoutes };
