--- api/src/certification/flash-certification/routes.js
@@ -1,5 +1,0 @@
-import * as scenarioSimulator from './application/scenario-simulator-route.js';
-
-const flashCertificationRoutes = [scenarioSimulator];
-
-export { flashCertificationRoutes };
