+++ api/src/shared/domain/models/index.js
@@ -1,7 +1,7 @@
 import { ComplementaryCertification } from '../../../certification/complementary-certification/domain/models/ComplementaryCertification.js';
+import { AssessmentSimulator } from '../../../certification/evaluation/domain/models/AssessmentSimulator.js';
+import { FlashAssessmentAlgorithm } from '../../../certification/evaluation/domain/models/FlashAssessmentAlgorithm.js';
 import { PartnerCertificationScoring } from '../../../certification/evaluation/domain/models/PartnerCertificationScoring.js';
-import { AssessmentSimulator } from '../../../certification/flash-certification/domain/models/AssessmentSimulator.js';
-import { FlashAssessmentAlgorithm } from '../../../certification/flash-certification/domain/models/FlashAssessmentAlgorithm.js';
 import { CertificationAttestation } from '../../../certification/results/domain/models/CertificationAttestation.js';
 import { PrivateCertificate } from '../../../certification/results/domain/models/PrivateCertificate.js';
 import { ResultCompetenceTree } from '../../../certification/results/domain/models/ResultCompetenceTree.js';
