+++ api/tests/certification/evaluation/unit/domain/models/FlashAssessmentAlgorithmPassageByAllCompetencesRule_test.js
@@ -1,4 +1,4 @@
-import { FlashAssessmentAlgorithmPassageByAllCompetencesRule } from '../../../../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithmPassageByAllCompetencesRule.js';
+import { FlashAssessmentAlgorithmPassageByAllCompetencesRule } from '../../../../../../src/certification/evaluation/domain/models/FlashAssessmentAlgorithmPassageByAllCompetencesRule.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | FlashAssessmentAlgorithm | FlashAssessmentAlgorithmPassageByAllCompetencesRule', function () {
