+++ api/tests/certification/evaluation/unit/infrastructure/serializers/scenario-simulator-batch-serializer_test.js
@@ -1,4 +1,4 @@
-import { scenarioSimulatorBatchSerializer } from '../../../../../../src/certification/flash-certification/infrastructure/serializers/scenario-simulator-batch-serializer.js';
+import { scenarioSimulatorBatchSerializer } from '../../../../../../src/certification/evaluation/infrastructure/serializers/scenario-simulator-batch-serializer.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | scenario-simulator-batch-serializer', function () {
