+++ api/tests/tooling/domain-builder/factory/certification/evaluation/build-flash-assessment-algorithm.js
@@ -1,0 +1,10 @@
+import { FlashAssessmentAlgorithm } from '../../../../../../src/certification/evaluation/domain/models/FlashAssessmentAlgorithm.js';
+import { FlashAssessmentAlgorithmRuleEngine } from '../../../../../../src/certification/evaluation/domain/models/FlashAssessmentAlgorithmRuleEngine.js';
+
+export const buildFlashAssessmentAlgorithm = ({ flashAlgorithmImplementation, configuration, ruleEngine }) => {
+  return new FlashAssessmentAlgorithm({
+    flashAlgorithmImplementation,
+    configuration,
+    ruleEngine: ruleEngine ?? new FlashAssessmentAlgorithmRuleEngine([], {}),
+  });
+};
