--- api/tests/tooling/domain-builder/factory/certification/flash-certification/build-flash-assessment-algorithm.js
@@ -1,10 +1,0 @@
-import { FlashAssessmentAlgorithm } from '../../../../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithm.js';
-import { FlashAssessmentAlgorithmRuleEngine } from '../../../../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithmRuleEngine.js';
-
-export const buildFlashAssessmentAlgorithm = ({ flashAlgorithmImplementation, configuration, ruleEngine }) => {
-  return new FlashAssessmentAlgorithm({
-    flashAlgorithmImplementation,
-    configuration,
-    ruleEngine: ruleEngine ?? new FlashAssessmentAlgorithmRuleEngine([], {}),
-  });
-};
